+++ packages/core/src/utilities/renderToCanvasCPU.ts
@@ -2,11 +2,13 @@ import {
   IImage,
   CPUFallbackEnabledElement,
   ViewportInputOptions,
+  IVolume,
 } from '../types';
 
 import getDefaultViewport from '../RenderingEngine/helpers/cpuFallback/rendering/getDefaultViewport';
 import calculateTransform from '../RenderingEngine/helpers/cpuFallback/rendering/calculateTransform';
 import drawImageSync from '../RenderingEngine/helpers/cpuFallback/drawImageSync';
+import type { CanvasLoadPosition } from './loadImageToCanvas';
 
 /**
  * Renders a cornerstone image object to a canvas.
@@ -17,11 +19,16 @@ import drawImageSync from '../RenderingEngine/helpers/cpuFallback/drawImageSync'
  */
 export default function renderToCanvasCPU(
   canvas: HTMLCanvasElement,
-  image: IImage,
+  imageOrVolume: IImage | IVolume,
   modality?: string,
   _renderingEngineId?: string,
   _viewportOptions?: ViewportInputOptions
-): Promise<string> {
+): Promise<CanvasLoadPosition> {
+  const volume = imageOrVolume as IVolume;
+  if (volume.volumeId) {
+    throw new Error('Unsupported volume rendering for CPU');
+  }
+  const image = imageOrVolume as IImage;
   const viewport = getDefaultViewport(canvas, image, modality);
 
   const enabledElement: CPUFallbackEnabledElement = {
@@ -36,6 +43,6 @@ export default function renderToCanvasCPU(
   const invalidated = true;
   return new Promise((resolve, reject) => {
     drawImageSync(enabledElement, invalidated);
-    resolve(image.imageId);
+    resolve(null);
   });
 }