+++ packages/tools/src/tools/annotation/BidirectionalTool.ts
@@ -158,6 +158,7 @@ class BidirectionalTool extends AnnotationTool {
         viewUp: <Types.Point3>[...viewUp],
         FrameOfReferenceUID,
         referencedImageId,
+        ...viewport.getViewReference({ points: [worldPos] }),
       },
       data: {
         handles: {