+++ packages/tools/src/tools/annotation/EllipticalROITool.ts
@@ -193,6 +193,7 @@ class EllipticalROITool extends AnnotationTool {
         viewUp: <Types.Point3>[...viewUp],
         FrameOfReferenceUID,
         referencedImageId,
+        ...viewport.getViewReference({ points: [worldPos] }),
       },
       data: {
         label: '',