+++ tests/volumeBasic.spec.js
@@ -6,18 +6,18 @@ import {
 } from './utils/index';
 
 test.beforeEach(async ({ page }) => {
-  await visitExample(page, 'stackBasic');
+  await visitExample(page, 'volumeBasic');
 });
 
-test.describe('Basic Stack', async () => {
-  test('should display a single DICOM image in a Stack viewport.', async ({
+test.describe('Basic Volume', async () => {
+  test('should display a single DICOM series in a Volume viewport.', async ({
     page,
   }) => {
     const locator = page.locator('.cornerstone-canvas');
     await checkForScreenshot(
       page,
       locator,
-      screenShotPaths.stackBasic.viewport
+      screenShotPaths.volumeBasic.viewport
     );
   });
 });