+++ packages/adapters/CHANGELOG.md
@@ -3,6 +3,66 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.77.2](https://github.com/dcmjs-org/dcmjs/compare/v1.77.1...v1.77.2) (2024-05-29)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
+## [1.77.1](https://github.com/dcmjs-org/dcmjs/compare/v1.77.0...v1.77.1) (2024-05-29)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
+# [1.77.0](https://github.com/dcmjs-org/dcmjs/compare/v1.76.1...v1.77.0) (2024-05-28)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
+## [1.76.1](https://github.com/dcmjs-org/dcmjs/compare/v1.76.0...v1.76.1) (2024-05-28)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
+# [1.76.0](https://github.com/dcmjs-org/dcmjs/compare/v1.75.2...v1.76.0) (2024-05-28)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
+## [1.75.2](https://github.com/dcmjs-org/dcmjs/compare/v1.75.1...v1.75.2) (2024-05-28)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
+## [1.75.1](https://github.com/dcmjs-org/dcmjs/compare/v1.75.0...v1.75.1) (2024-05-28)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
+# [1.75.0](https://github.com/dcmjs-org/dcmjs/compare/v1.74.8...v1.75.0) (2024-05-28)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
+## [1.74.8](https://github.com/dcmjs-org/dcmjs/compare/v1.74.7...v1.74.8) (2024-05-24)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
+## [1.74.7](https://github.com/dcmjs-org/dcmjs/compare/v1.74.6...v1.74.7) (2024-05-23)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
+## [1.74.6](https://github.com/dcmjs-org/dcmjs/compare/v1.74.5...v1.74.6) (2024-05-23)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
+## [1.74.5](https://github.com/dcmjs-org/dcmjs/compare/v1.74.4...v1.74.5) (2024-05-22)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
+## [1.74.4](https://github.com/dcmjs-org/dcmjs/compare/v1.74.3...v1.74.4) (2024-05-21)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
+## [1.74.3](https://github.com/dcmjs-org/dcmjs/compare/v1.74.2...v1.74.3) (2024-05-21)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
+## [1.74.2](https://github.com/dcmjs-org/dcmjs/compare/v1.74.1...v1.74.2) (2024-05-21)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
 ## [1.74.1](https://github.com/dcmjs-org/dcmjs/compare/v1.74.0...v1.74.1) (2024-05-17)
 
 ### Bug Fixes